% Generated by build script
\name{figshare_collections}
\alias{figshare_collections}
\title{Search Figshare collections}
\usage{
figshare_collections()
}
\arguments{
\item{query}{Search query string}
\item{page_size}{Number of results (default 10)}
\item{page}{Page number (default 1)}
}
\value{tibble: id, title, doi, url, published_date, articles_count}
\description{Search Figshare collections}
