% Generated by build script
\name{figshare_files}
\alias{figshare_files}
\title{List files attached to a Figshare article}
\usage{
figshare_files()
}
\arguments{
\item{id}{Figshare article ID (integer)}
}
\value{tibble: id, name, size, download_url, computed_md5}
\description{List files attached to a Figshare article}
